Use ivy-like for files
So fuzzy matching (like with `flx-ido`) across files in large repos is
slow (whether with ido or ivy), instead ivy-like matching works great
for files. But with ivy, we can keep fuzzy matching for commands and
buffers where is fast and useful.

Also `C-o m` can be used to switch matchers.