Replace `counsel-rg` with my own `rg-ask-dwim`
The awful thing about `rg-dwim` was that it was the only command which
didn't ask for a file type. I almost always want to use
"everything" (that is, don't filter on file types), but `rg-dwim` didn't
ask for query. The funny thing is that the other functions still
defaulted to searching for the thing-at-point, just with a prompt. I
almost replaced this package with `deadgrep`, but I like that it's based
on `compilation-mode`. Fortunately I was able to define my own search
function which asks for the query, but otherwise searches everything in
my project (just like how I use ripgrep on the command-line).