Replace orderless with built-in completion-styles
With the utility of partial-completion and the addition of initials and
flex as fallback styles, and history seemingly "just working now," we
the built-in styles work just fine.

So initials and partial-completion are hella cool because `~/sel`
matches `~/src/emacs/lisp` (and the latter just does the same but beyond
initials), negating the need for `/` to complete a directory. Still
useful to have `SPC` bound to that though.